+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 1. add `snippets.sql.SqliteSQL` as Sqlite SQL generator
 2. add `cmd.parse_deps` to parse dependencies of a project directory, and find circular dependencies.
 3. add `cmd.os.linux.systemd.service` and `cmd.os.linux.systemd.timer`
+4. add `functools.to_thread`, same as `asyncio.to_thread` in python 3.9+.
 
 ### 1.1.8 (2024-12-11)
 1. add `utils.i2b` and `utils.b2i`, integer and fixed-length byte strings conversion.